requirements.txt
asgiref==3.2.10
boto3==1.14.43
botocore==1.17.43
cachetools==4.1.1
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.3
django-admin-interface==0.12.2
django-ckeditor==5.9.0
django-colorfield==0.3.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.2.2
django-storages==1.9.1
django-taggit==1.3.0
docopt==0.6.2
docutils==0.15.2
google-api-core==1.22.1
google-auth==1.20.1
google-cloud-core==1.4.1
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
gunicorn==20.0.4
html5lib==1.1
idna==2.10
jmespath==0.10.0
num2words==0.5.10
Pillow==7.2.0
protobuf==3.13.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pyphen==0.9.5
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
requests==2.24.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
setuptools==40.3.0
tinycss2==1.0.2
urllib3==1.25.10
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.2.0
